requirements.txt
streamlit
authlib
python-dotenv
